sr(api): move attach-target-profiles-to-training in devcomp
